use new core go module
